+++ ppdet/data/transform/operators.py
@@ -1712,7 +1712,7 @@ class Mixup(BaseOperator):
             gt_score2 = np.ones_like(sample[1]['gt_class'])
             gt_score = np.concatenate(
                 (gt_score1 * factor, gt_score2 * (1. - factor)), axis=0)
-            result['gt_score'] = gt_score
+            result['gt_score'] = gt_score.astype('float32')
         if 'is_crowd' in sample[0]:
             is_crowd1 = sample[0]['is_crowd']
             is_crowd2 = sample[1]['is_crowd']
@@ -2604,7 +2604,7 @@ class Mosaic(BaseOperator):
             difficult = np.concatenate([x['difficult'] for x in sample], axis=0)
         sample = sample[0]
         sample['image'] = image.astype(np.uint8)
-        sample['gt_bbox'] = boxes
+        sample['gt_bbox'] = boxes.astype(np.float32)
         sample['gt_class'] = labels
         if 'is_crowd' in sample:
             sample['is_crowd'] = is_crowd
